+++ src/lsts.h
@@ -47,7 +47,7 @@ void blat(struct BList *, uc *, uint32_t);
 #define blist_clear(l) ((l)->size = 0)
 void blist_clear_free(struct BList *);
 void blist_print(struct BList *);
-void blist_add_set(struct BList *, uc, uint32_t *, size_t);
+void blist_add_set(struct BList *, uc, long *, size_t);
 struct BList *copy_str(struct BList *src);
 struct BList *copy_blist(struct BList *l);
 #define b_last(l) (blist_get((l), (l)->size - 1))
